+++ lib/Api/TransactionsApi.php
@@ -114,7 +114,7 @@ class TransactionsApi
             $headerParams['Accept'] = $_header_accept;
         }
         $headerParams['Content-Type'] = ApiClient::selectHeaderContentType(array('application/json'));
-        $headerParams['Square-Version'] = "2019-03-27";
+        $headerParams['Square-Version'] = "2019-04-10";
 
         
         
@@ -223,7 +223,7 @@ class TransactionsApi
             $headerParams['Accept'] = $_header_accept;
         }
         $headerParams['Content-Type'] = ApiClient::selectHeaderContentType(array('application/json'));
-        $headerParams['Square-Version'] = "2019-03-27";
+        $headerParams['Square-Version'] = "2019-04-10";
 
         
         
@@ -335,7 +335,7 @@ class TransactionsApi
             $headerParams['Accept'] = $_header_accept;
         }
         $headerParams['Content-Type'] = ApiClient::selectHeaderContentType(array('application/json'));
-        $headerParams['Square-Version'] = "2019-03-27";
+        $headerParams['Square-Version'] = "2019-04-10";
 
         
         
@@ -450,7 +450,7 @@ class TransactionsApi
             $headerParams['Accept'] = $_header_accept;
         }
         $headerParams['Content-Type'] = ApiClient::selectHeaderContentType(array('application/json'));
-        $headerParams['Square-Version'] = "2019-03-27";
+        $headerParams['Square-Version'] = "2019-04-10";
 
         // query params
         if ($begin_time !== null) {
@@ -566,7 +566,7 @@ class TransactionsApi
             $headerParams['Accept'] = $_header_accept;
         }
         $headerParams['Content-Type'] = ApiClient::selectHeaderContentType(array('application/json'));
-        $headerParams['Square-Version'] = "2019-03-27";
+        $headerParams['Square-Version'] = "2019-04-10";
 
         // query params
         if ($begin_time !== null) {
@@ -680,7 +680,7 @@ class TransactionsApi
             $headerParams['Accept'] = $_header_accept;
         }
         $headerParams['Content-Type'] = ApiClient::selectHeaderContentType(array('application/json'));
-        $headerParams['Square-Version'] = "2019-03-27";
+        $headerParams['Square-Version'] = "2019-04-10";
 
         
         
@@ -789,7 +789,7 @@ class TransactionsApi
             $headerParams['Accept'] = $_header_accept;
         }
         $headerParams['Content-Type'] = ApiClient::selectHeaderContentType(array('application/json'));
-        $headerParams['Square-Version'] = "2019-03-27";
+        $headerParams['Square-Version'] = "2019-04-10";