+++ lib/Model/V1Discount.php
@@ -30,7 +30,8 @@ class V1Discount implements ArrayAccess
         'amount_money' => '\SquareConnect\Model\V1Money',
         'discount_type' => 'string',
         'pin_required' => 'bool',
-        'color' => 'string'
+        'color' => 'string',
+        'v2_id' => 'string'
     );
   
     /** 
@@ -44,7 +45,8 @@ class V1Discount implements ArrayAccess
         'amount_money' => 'amount_money',
         'discount_type' => 'discount_type',
         'pin_required' => 'pin_required',
-        'color' => 'color'
+        'color' => 'color',
+        'v2_id' => 'v2_id'
     );
   
     /**
@@ -58,7 +60,8 @@ class V1Discount implements ArrayAccess
         'amount_money' => 'setAmountMoney',
         'discount_type' => 'setDiscountType',
         'pin_required' => 'setPinRequired',
-        'color' => 'setColor'
+        'color' => 'setColor',
+        'v2_id' => 'setV2Id'
     );
   
     /**
@@ -72,7 +75,8 @@ class V1Discount implements ArrayAccess
         'amount_money' => 'getAmountMoney',
         'discount_type' => 'getDiscountType',
         'pin_required' => 'getPinRequired',
-        'color' => 'getColor'
+        'color' => 'getColor',
+        'v2_id' => 'getV2Id'
     );
   
     /**
@@ -110,6 +114,11 @@ class V1Discount implements ArrayAccess
       * @var string
       */
     protected $color;
+    /**
+      * $v2_id The ID of the CatalogObject in the Connect v2 API. Objects that are shared across multiple locations share the same v2 ID.
+      * @var string
+      */
+    protected $v2_id;
 
     /**
      * Constructor
@@ -152,6 +161,11 @@ class V1Discount implements ArrayAccess
               $this->color = $data["color"];
             } else {
               $this->color = null;
+            }
+            if (isset($data["v2_id"])) {
+              $this->v2_id = $data["v2_id"];
+            } else {
+              $this->v2_id = null;
             }
         }
     }
@@ -286,6 +300,25 @@ class V1Discount implements ArrayAccess
     public function setColor($color)
     {
         $this->color = $color;
+        return $this;
+    }
+    /**
+     * Gets v2_id
+     * @return string
+     */
+    public function getV2Id()
+    {
+        return $this->v2_id;
+    }
+  
+    /**
+     * Sets v2_id
+     * @param string $v2_id The ID of the CatalogObject in the Connect v2 API. Objects that are shared across multiple locations share the same v2 ID.
+     * @return $this
+     */
+    public function setV2Id($v2_id)
+    {
+        $this->v2_id = $v2_id;
         return $this;
     }
     /**