+++ lib/Model/V1ModifierList.php
@@ -27,7 +27,8 @@ class V1ModifierList implements ArrayAccess
         'id' => 'string',
         'name' => 'string',
         'selection_type' => 'string',
-        'modifier_options' => '\SquareConnect\Model\V1ModifierOption[]'
+        'modifier_options' => '\SquareConnect\Model\V1ModifierOption[]',
+        'v2_id' => 'string'
     );
   
     /** 
@@ -38,7 +39,8 @@ class V1ModifierList implements ArrayAccess
         'id' => 'id',
         'name' => 'name',
         'selection_type' => 'selection_type',
-        'modifier_options' => 'modifier_options'
+        'modifier_options' => 'modifier_options',
+        'v2_id' => 'v2_id'
     );
   
     /**
@@ -49,7 +51,8 @@ class V1ModifierList implements ArrayAccess
         'id' => 'setId',
         'name' => 'setName',
         'selection_type' => 'setSelectionType',
-        'modifier_options' => 'setModifierOptions'
+        'modifier_options' => 'setModifierOptions',
+        'v2_id' => 'setV2Id'
     );
   
     /**
@@ -60,7 +63,8 @@ class V1ModifierList implements ArrayAccess
         'id' => 'getId',
         'name' => 'getName',
         'selection_type' => 'getSelectionType',
-        'modifier_options' => 'getModifierOptions'
+        'modifier_options' => 'getModifierOptions',
+        'v2_id' => 'getV2Id'
     );
   
     /**
@@ -83,6 +87,11 @@ class V1ModifierList implements ArrayAccess
       * @var \SquareConnect\Model\V1ModifierOption[]
       */
     protected $modifier_options;
+    /**
+      * $v2_id The ID of the CatalogObject in the Connect v2 API. Objects that are shared across multiple locations share the same v2 ID.
+      * @var string
+      */
+    protected $v2_id;
 
     /**
      * Constructor
@@ -110,6 +119,11 @@ class V1ModifierList implements ArrayAccess
               $this->modifier_options = $data["modifier_options"];
             } else {
               $this->modifier_options = null;
+            }
+            if (isset($data["v2_id"])) {
+              $this->v2_id = $data["v2_id"];
+            } else {
+              $this->v2_id = null;
             }
         }
     }
@@ -187,6 +201,25 @@ class V1ModifierList implements ArrayAccess
     public function setModifierOptions($modifier_options)
     {
         $this->modifier_options = $modifier_options;
+        return $this;
+    }
+    /**
+     * Gets v2_id
+     * @return string
+     */
+    public function getV2Id()
+    {
+        return $this->v2_id;
+    }
+  
+    /**
+     * Sets v2_id
+     * @param string $v2_id The ID of the CatalogObject in the Connect v2 API. Objects that are shared across multiple locations share the same v2 ID.
+     * @return $this
+     */
+    public function setV2Id($v2_id)
+    {
+        $this->v2_id = $v2_id;
         return $this;
     }
     /**