+++ lib/Model/V1ModifierOption.php
@@ -29,7 +29,8 @@ class V1ModifierOption implements ArrayAccess
         'price_money' => '\SquareConnect\Model\V1Money',
         'on_by_default' => 'bool',
         'ordinal' => 'int',
-        'modifier_list_id' => 'string'
+        'modifier_list_id' => 'string',
+        'v2_id' => 'string'
     );
   
     /** 
@@ -42,7 +43,8 @@ class V1ModifierOption implements ArrayAccess
         'price_money' => 'price_money',
         'on_by_default' => 'on_by_default',
         'ordinal' => 'ordinal',
-        'modifier_list_id' => 'modifier_list_id'
+        'modifier_list_id' => 'modifier_list_id',
+        'v2_id' => 'v2_id'
     );
   
     /**
@@ -55,7 +57,8 @@ class V1ModifierOption implements ArrayAccess
         'price_money' => 'setPriceMoney',
         'on_by_default' => 'setOnByDefault',
         'ordinal' => 'setOrdinal',
-        'modifier_list_id' => 'setModifierListId'
+        'modifier_list_id' => 'setModifierListId',
+        'v2_id' => 'setV2Id'
     );
   
     /**
@@ -68,7 +71,8 @@ class V1ModifierOption implements ArrayAccess
         'price_money' => 'getPriceMoney',
         'on_by_default' => 'getOnByDefault',
         'ordinal' => 'getOrdinal',
-        'modifier_list_id' => 'getModifierListId'
+        'modifier_list_id' => 'getModifierListId',
+        'v2_id' => 'getV2Id'
     );
   
     /**
@@ -101,6 +105,11 @@ class V1ModifierOption implements ArrayAccess
       * @var string
       */
     protected $modifier_list_id;
+    /**
+      * $v2_id The ID of the CatalogObject in the Connect v2 API. Objects that are shared across multiple locations share the same v2 ID.
+      * @var string
+      */
+    protected $v2_id;
 
     /**
      * Constructor
@@ -138,6 +147,11 @@ class V1ModifierOption implements ArrayAccess
               $this->modifier_list_id = $data["modifier_list_id"];
             } else {
               $this->modifier_list_id = null;
+            }
+            if (isset($data["v2_id"])) {
+              $this->v2_id = $data["v2_id"];
+            } else {
+              $this->v2_id = null;
             }
         }
     }
@@ -253,6 +267,25 @@ class V1ModifierOption implements ArrayAccess
     public function setModifierListId($modifier_list_id)
     {
         $this->modifier_list_id = $modifier_list_id;
+        return $this;
+    }
+    /**
+     * Gets v2_id
+     * @return string
+     */
+    public function getV2Id()
+    {
+        return $this->v2_id;
+    }
+  
+    /**
+     * Sets v2_id
+     * @param string $v2_id The ID of the CatalogObject in the Connect v2 API. Objects that are shared across multiple locations share the same v2 ID.
+     * @return $this
+     */
+    public function setV2Id($v2_id)
+    {
+        $this->v2_id = $v2_id;
         return $this;
     }
     /**